:recycle: move business code outside of data model package
